// src/app/job-list/job.interface.ts
export interface Job {
    Tyono: number;
    Paivamara: Date;
    Toimituspaiva: Date;
    Tekija: string;
    Tyo: string;
    Asiakas: string;
    Yritys: string;
    Laskutusosoite: string;
    Viite: number;
    Tarjottu: number;
    Kustannusarvio: number;
    Koko: string;
    Sivumaara: number;
    Painosmaara: number;
    Varit: string;
    Paperi: string;
    Jalkikasittely: string;
    Toimkulut: number;
    Huom: string;
    Tunnit: number;
    Laskutusnumero: number;
}